+++ CHANGELOG.md
@@ -4,6 +4,9 @@ NCS Navigator MDES Warehouse History
 0.3.2
 -----
 
+- Correct VDR XML reader for the case where an MDES table that
+  contains a variable with the same name as the table. (#1695)
+
 0.3.1
 -----